Migrate ActivityHeatmap to Vue SFC
